Changed Rotation/Yaw to be Degrees everywhere instead of bytes or Quaternions in different parts. Fixed up mapLinks and moved around some NPCs

+++ AISpace.Common/DAL/Entities/MapLink.cs
@@ -20,7 +20,10 @@ public class MapLink
     public float PositionX { get; set; }
     public float PositionY { get; set; }
     public float PositionZ { get; set; }
-    public byte Yaw { get; set; }
+
+    /// <summary>Trigger orientation in degrees.</summary>
+    public int Yaw { get; set; }
+
     public float Length { get; set; }
     public float Depth { get; set; }
